+++ packages/create-app-shared/src/miniapp/getSearchParams.ts
@@ -1,5 +1,5 @@
 import { getHistory } from '../storage';
 
 export default function(history = getHistory()) {
-  return (history.location as any)._currentPageOptions;
+  return (history?.location as any)?._currentPageOptions || {};
 }