+++ packages/plugin-app-core/src/index.ts
@@ -8,15 +8,12 @@ import { setAlias, setProjectType, setEntry, setTempDir, setRegisterMethod, setR
 import getBuildConfig from './utils/getBuildConfig';
 
 // eslint-disable-next-line
-const chalk = require('chalk');
-// eslint-disable-next-line
 const { constants: { MINIAPP, WECHAT_MINIPROGRAM, BAIDU_SMARTPROGRAM, KUAISHOU_MINIPROGRAM, QUICKAPP, BYTEDANCE_MICROAPP } } = require('miniapp-builder-shared');
 const miniappPlatforms = [ MINIAPP, WECHAT_MINIPROGRAM, BYTEDANCE_MICROAPP, BAIDU_SMARTPROGRAM, KUAISHOU_MINIPROGRAM ];
 
 export default (api, options) => {
   const { onHook, context, setValue } = api;
   const { command, userConfig, rootDir } = context;
-  const { targets = ['web'] } = userConfig;
   const { framework } = options;
 
   // Set framework field
@@ -43,9 +40,6 @@ export default (api, options) => {
   // Set webpack cache id
   setValue('WEBPACK_CACHE_ID', hash({ ...userConfig, hasJsxRuntime }));
 
-  // Check target
-  checkTargets(targets);
-
   // Set temporary directory
   // eg: .ice or .rax
   setTempDir(api, options);
@@ -66,6 +60,7 @@ export default (api, options) => {
 
   // register api method
   const generator = initGenerator(api, { ...options, hasJsxRuntime });
+
   setRegisterMethod(api, { generator });
 
   // add core template for framework
@@ -140,6 +135,8 @@ function getDefaultRenderData(api, options) {
       isMPA: false,
       tabBarPath: '', // avoid ejs error
       routesFilePath: './staticConfig',
+      // MPA 下会覆盖
+      enableRouter: true,
     };
   } else {
     return {
@@ -147,45 +144,8 @@ function getDefaultRenderData(api, options) {
       isReact: true,
       isRax: false,
       ssr,
+      // MPA 下会覆盖
+      enableRouter: (!userConfig.mpa && userConfig.router !== false),
     };
   }
-}
-
-function checkTargets(targets) {
-  let hasError = false;
-
-  if (Object.prototype.toString.call(targets) === '[object Object]') {
-    hasError = true;
-  }
-
-  if (typeof targets === 'string') {
-    hasError = true;
-  }
-
-  if (Array.isArray(targets) && !matchTargets(targets)) {
-    hasError = true;
-  }
-
-  if (hasError) {
-    const msg = `
-  targets must be the array type in build.json.
-
-    e.g. { "targets": ["miniapp", "wechat-miniprogram"] }
-
-  if you want to describes the browserslist environments for your project.
-  you should set browserslist in build.json.
-
-    e.g. { "browserslist": { "chrome": "58", "ie": 11 } }
-`;
-    console.log();
-    console.log(chalk.red(msg));
-    console.log();
-    process.exit(1);
-  }
-}
-
-function matchTargets(targets) {
-  return targets.every(target => {
-    return ['web', 'weex', 'kraken', MINIAPP, WECHAT_MINIPROGRAM, BYTEDANCE_MICROAPP, BAIDU_SMARTPROGRAM, KUAISHOU_MINIPROGRAM, QUICKAPP].includes(target);
-  });
 }