+++ packages/plugin-app-core/src/runtime.tsx
@@ -1,7 +1,0 @@
-const module = ({ addProvider, appConfig }) => {
-  if (appConfig.app && appConfig.app.addProvider) {
-    addProvider(appConfig.app.addProvider);
-  }
-};
-
-export default module;