+++ packages/plugin-icestark/src/runtime.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { useEffect, useState } from 'react';
+import { useEffect, useState, useMemo, memo } from 'react';
 import * as ReactDOM from 'react-dom';
 import { AppConfig, AppRouter, AppRoute } from '@ice/stark';
 import {
@@ -14,7 +14,7 @@ import { IceRouter } from '$ice/Router';
 // @ts-ignore
 import DefaultLayout from '$ice/Layout';
 import removeRootLayout from './runtime/removeLayout';
-import { IPrivateIceStark, IIceStark } from './types';
+import { IPrivateIceStark, IIceStark, IStarkAppConfig } from './types';
 
 const module = ({
   appConfig,
@@ -30,7 +30,13 @@ const module = ({
   wrapperRouteComponent
 }) => {
   const { icestark, router } = appConfig;
-  const { type: appType = process.env.__ICESTARK_TYPE__, registerAppEnter: enterRegistration, registerAppLeave: leaveRegistration, $$props } = (icestark || {}) as IPrivateIceStark;
+  const {
+    type: appType = process.env.__ICESTARK_TYPE__,
+    registerAppEnter: enterRegistration,
+    registerAppLeave: leaveRegistration,
+    getApps,
+    $$props
+  } = (icestark || {}) as IPrivateIceStark;
   const { type, basename, modifyRoutes: runtimeModifyRoutes, fallback } = router;
   // compatible with deprecated runtime API
   const wrapperComponent = wrapperPageComponent || wrapperRouteComponent;
@@ -54,32 +60,40 @@ const module = ({
       return new Promise(resolve => {
         if (isInIcestark()) {
           if (localIcestarkType === 'normal') {
-            registerAppEnter(() => {
-              const mountNode = getMountNode();
+            // @ts-ignore remove this next time for https://github.com/ice-lab/icestark/pull/440
+            registerAppEnter((props) => {
+              const container = (props && props.container) || getMountNode();
               if (enterRegistration) {
-                enterRegistration(mountNode, App, resolve);
+                enterRegistration(container, App, resolve);
               } else {
-                ReactDOM.render(<App />, mountNode, resolve);
+                ReactDOM.render(<App />, container, () => {
+                  resolve(true);
+                });
               }
             });
             // make sure the unmount event is triggered
-            registerAppLeave(() => {
-              const mountNode = getMountNode();
+            // @ts-ignore remove this next time for https://github.com/ice-lab/icestark/pull/440
+            registerAppLeave((props) => {
+              const container = (props && props.container) || getMountNode();
               if (leaveRegistration) {
-                leaveRegistration(mountNode);
+                leaveRegistration(container);
               } else {
-                ReactDOM.unmountComponentAtNode(mountNode);
+                ReactDOM.unmountComponentAtNode(container);
               }
             });
           } else {
             let { container } = $$props ?? {};
             if (!container) {
-              container = getMountNode();
+              container = getMountNode() as HTMLElement;
             }
-            ReactDOM.render(<App />, container, resolve);
+            ReactDOM.render(<App />, container, () => {
+              resolve(true);
+            });
           }
         } else {
-          ReactDOM.render(<App />, appMountNode, resolve);
+          ReactDOM.render(<App />, appMountNode, () => {
+            resolve(true);
+          });
         }
       });
     });
@@ -115,8 +129,9 @@ const module = ({
         return <IceRouter {...routerProps} routes={routes} />;
       });
     }
-  } else if (appType === 'framework') {
-    const { getApps, appRouter, Layout, AppRoute: CustomAppRoute, removeRoutesLayout } = (icestark || {}) as IIceStark;
+  } else if (appType === 'framework' && getApps) {
+    const { appRouter, Layout, AppRoute: CustomAppRoute, removeRoutesLayout } = (icestark || {}) as IIceStark;
+
     if (removeRoutesLayout) {
       modifyRoutes(removeRootLayout);
     }
@@ -133,12 +148,12 @@ const module = ({
     };
 
     const frameworkRouter = (routes) => () => {
-      const [appPathname, setAppPathname] = useState('');
+      const [appPathname, setAppPathname] = useState(window.location.pathname);
       const [routeInfo, setRouteInfo] = useState({});
       const [appEnter, setAppEnter] = useState({});
       const [appLeave, setAppLeave] = useState({});
 
-      const [apps, setApps] = useState(null);
+      const [apps, setApps] = useState<IStarkAppConfig[] | null>(null);
       const BasicLayout = Layout || DefaultLayout || ((props) => (<>{props.children}</>));
       const RenderAppRoute = (CustomAppRoute || AppRoute) as typeof AppRoute;
 
@@ -171,6 +186,12 @@ const module = ({
         updateApps: setApps,
       };
 
+      // RootApp will re-render on every AppRoute's update if RootApp were matched.
+      const MemoRootApp = useMemo(
+        () => memo(() => <RootApp routes={routes} />),
+        []
+      );
+
       return (
         <BasicLayout {...appInfo}>
           {apps && (
@@ -192,7 +213,7 @@ const module = ({
                 <RenderAppRoute
                   path="/"
                   render={() => {
-                    return <RootApp routes={routes} />;
+                    return <MemoRootApp />;
                   }}
                 />
               )}
@@ -202,6 +223,13 @@ const module = ({
       );
     };
     setRenderComponent(frameworkRouter);
+  }
+
+  if (appType === 'framework' && !getApps) {
+    console.warn(`
+      [plugin-icestark]: appConfig.icestark.getApps should be not empty if this is an framework app; If not，please make sure appConfgi.icestark.type exist.
+      see https://ice.work/docs/guide/advanced/icestark/
+    `);
   }
 };