+++ packages/react-app-renderer/src/types.ts
@@ -6,7 +6,8 @@ export type OnError = (err: Error, componentStack: string) => void
 export interface Context {
   initialContext: InitialContext,
   initialData: { [k: string]: any },
-  pageInitialProps: { [k: string]: any }
+  pageInitialProps: { [k: string]: any },
+  enableRouter?: boolean,
 }
 
 export interface ServerContext {
@@ -37,8 +38,8 @@ export type RenderAppConfig = {
     ErrorBoundaryFallback?: React.ComponentType;
     errorBoundary?: boolean;
     getInitialData?: (context: InitialContext) => Promise<any>;
+    renderComponent?: React.ComponentType;
   },
-  renderComponent?: React.ComponentType
 };
 export type AppLifecycle = {
   createBaseApp: <T>(appConfig: T, buildConfig: any, context: any) => { runtime: RuntimeModule; appConfig: T };