+++ test/utils/build.ts
@@ -25,6 +25,7 @@ export const buildFixture = function(example: string) {
     const processCwdSpy = jest.spyOn(process, 'cwd');
     processCwdSpy.mockReturnValue(rootDir);
     process.env.DISABLE_FS_CACHE = 'true';
+
     await build({
       args: {
         config: path.join(rootDir, 'build.json'),