+++ packages/build-jest-config/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.0.0
+
+- [fix] bump version of `@builder/pack`
+
 ## 1.0.1
 
 - [fix] export as commonjs for babelTransform