+++ packages/build-webpack-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # changelog
 
+## 2.0.0
+
+- [fix] bump version of `@builder/pack`(^0.5.0)
+
+## 1.0.4
+
+- [fix] lock postcss version for postcss-loader
+
+## 1.0.3
+
+- [feat] add postcss-nested for css nesting rules
+
 ## 1.0.2
 
 - [fix] bump version of `@builder/pack`