+++ packages/plugin-app-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 2.1.0
+
+- [feat] register `router`&`request` to userConfig
+- [chore] pass `enableRouter` to generator
+- [feat] remove runtime addProvider and move to react-app/rax-app
+- [feat] add `mpa` to runtime buildConfig
+- [fix] update dependencies version
+
+## 2.0.3
+
+- [feat] format rax plugin name
+
 ## 2.0.2
 
 - [feat] support load `styl` style