+++ packages/plugin-config/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.0.2
+
+- [fix] polyfill for globalThis
+
+## 2.0.1
+
+- [fix] generic conflict with JSX
+
 ## 2.0.0
 
 - [feat] migrate runtime api