+++ packages/plugin-react-app/CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## 2.1.2
+
+- [feat] feat: support `props.pageConfig` for each page component
+- [feat] pre bundle `webpack-dev-server`
+- [fix] bump version of `@builder/pack`(^0.5.0)
+- [fix] bump version of `@builder/user-config`, `query-loader-webpack-plugin`, `@builder/webpack-config`
+
+## 2.1.1
+
+- [fix] avoid error when disableRuntime
+
+## 2.1.0
+
+- [feat] add runtime abilities: addProvider/WrapperErrorBoundary/WrapperCSR/WrapperSSR/WrapperSearchParams (move from plugin-router)
+- [feat] add runtime router options check and log warning
+- [fix] update dependencies version
+
+## 2.0.4
+
+- [fix] error occurred when config externals and remoteRuntime
+- [fix] build error when active `remoteRuntime`
+
+## 2.0.3
+
+- [feat] support apply static loader with query like `?worker|?url|?raw|?worker&inline`
+- [fix] typo of `vitePlugins`
+- [fix] bump `esbuild` version to `^0.13.12`
+
 ## 2.0.2
 
 - [fix] bump version of `webpack-plugin-import`