+++ packages/plugin-store/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 2.0.6
+
+- [fix] update dependencies version
+
+## 2.0.5
+
+- [fix] remove unnecessary import redirection
+
+## 2.0.4
+
+- [fix] store error in vite HMR
+
 ## 2.0.3
 
 - [fix] nested routes not render in vite mode